restructuradas interfaces para estructura del sistema.
