+++ src/app/Repositories/Interfaces/Aux/CategoryRepositoryInterface.php
@@ -1,5 +1,7 @@
-<?php namespace PCI\Repositories\Interfaces;
+<?php namespace PCI\Repositories\Interfaces\Aux;
 
+use PCI\Repositories\Interfaces\GetBySlugOrIdInterface;
+use PCI\Repositories\Interfaces\ModelRepositoryInterface;
 use PCI\Repositories\Interfaces\Viewable\ViewableInterface;
 
 interface CategoryRepositoryInterface extends ModelRepositoryInterface, ViewableInterface, GetBySlugOrIdInterface
